Refactor abstract classes
